Control.Iterate.SetAlgebra

Documentation

class Iter f ⇒ Basic f where

In order to build typed Exp (which are a typed deep embedding) of Set operations, we need to know what kind of basic types of Maps and Sets can be embedded. Every Basic type has a few operations for creating one from a list, for adding and removing key-value pairs, looking up a value given a key. Instances of this algebra are functional in that every key has exactly one value associated with it. ===================================================================================================

Minimal complete definition

addkv, removekey, domain, range

Methods

addpair ∷ Ord k ⇒ k → v → f k v → f k v

in addpair the new value always prevails, to make a choice use addkv which has a combining function that allows choice.

addkv ∷ Ord k ⇒ (k, v) → f k v → (v → v → v) → f k v

use ( old new -> old) if you want the v in (f k v) to prevail, and use ( old new -> new) if you want the v in (k,v) to prevail

removekey ∷ Ord k ⇒ k → f k v → f k v

domain ∷ Ord k ⇒ f k v → Set k

range ∷ Ord v ⇒ f k v → Set v

emptyc ∷ Ord k ⇒ f k v

fromPairs ∷ Ord k ⇒ (v → v → v) → [(k, v)] → List k v

normalize ∷ Ord k ⇒ (v → v → v) → [(k, v)] → [(k, v)]

data Single k v where

Constructors

Single ∷ k → v → Single k v

Fail ∷ Single k v

SetSingle ∷ k → Single k ()

mapflip ∷ (v → v → v) → v → v → v

data BiMap v a b where

Constructors

MkBiMap ∷ v ~ b ⇒ !(Map a b) → !(Map b (Set a)) → BiMap v a b

addBack ∷ (Ord v, Ord k) ⇒ v → k → Map v (Set k) → Map v (Set k)

retract ∷ (Ord v, Ord k) ⇒ v → k → Map v (Set k) → Map v (Set k)

insertBackwards ∷ (Ord k, Ord v) ⇒ v → v → k → Map v (Set k) → Map v (Set k)

biMapEmpty ∷ BiMap v k v

biMapFromList ∷ (Ord k, Ord v) ⇒ (v → v → v) → [(k, v)] → BiMap v k v

type Bimap k v = BiMap v k v

removeval ∷ (Ord k, Ord v) ⇒ v → BiMap v k v → BiMap v k v

data Sett k v where

Constructors

Sett ∷ Set k → Sett k ()

class Embed concrete base | concrete → base where

Methods

toBase ∷ concrete → base

fromBase ∷ base → concrete

class Iter f where

Minimal complete definition

nxt, lub

Methods

nxt ∷ f a b → Collect (a, b, f a b)

lub ∷ Ord k ⇒ k → f k b → Collect (k, b, f k b)

hasNxt ∷ f a b → Maybe (a, b, f a b)

hasLub ∷ Ord k ⇒ k → f k b → Maybe (k, b, f k b)

haskey ∷ Ord key ⇒ key → f key b → Bool

isnull ∷ f k v → Bool

lookup ∷ Ord key ⇒ key → f key rng → Maybe rng

element ∷ Ord k ⇒ k → f k v → Collect ()

data List k v where

Constructors

UnSafeList ∷ Ord k ⇒ [(k, v)] → List k v

unList ∷ List k v → [(k, v)]

noKeys ∷ Ord k ⇒ Map k a → Map k b → Map k a

keysEqual ∷ Ord k ⇒ Map k v1 → Map k v2 → Bool

sameDomain ∷ (Ord k, Iter f, Iter g) ⇒ f k b → g k c → Bool

splitMember ∷ Ord k ⇒ k → Map k a → (Map k a, Bool, Map k a)

A variant of splitLookup that indicates only whether the key was present, rather than producing its value. This is used to implement keysEqual to avoid allocating unnecessary Just constructors.

data StrictTriple a b c

Constructors

StrictTriple !a !b !c

intersectDomP ∷ Ord k ⇒ (k → v2 → Bool) → Map k v1 → Map k v2 → Map k v2

intersetDomP p m1 m2 == Keep the key and value from m2, iff (the key is in the dom of m1) && ((p key value) is true)

intersectDomPLeft ∷ Ord k ⇒ (k → v2 → Bool) → Map k v1 → Map k v2 → Map k v1

Similar to intersectDomP, except the Map returned has the same key as the first input map, rather than the second input map.

intersectMapSetFold ∷ Ord k ⇒ (k → v → ans → ans) → Map k v → Set k → ans → ans

fold over the intersection of a Map and a Set

disjointMapSetFold ∷ Ord k ⇒ (k → v → ans → ans) → Map k v → Set k → ans → ans

Fold with accum all those pairs in the map, not appearing in the set.

data BaseRep f k v where

Constructors

MapR ∷ Basic Map ⇒ BaseRep Map k v

SetR ∷ Basic Sett ⇒ BaseRep Sett k ()

ListR ∷ Basic List ⇒ BaseRep List k v

SingleR ∷ Basic Single ⇒ BaseRep Single k v

BiMapR ∷ (Basic (BiMap v), Ord v) ⇒ BaseRep (BiMap v) k v

lifo ∷ Iter f ⇒ f k v → Collect (k, v)

fifo ∷ Iter f ⇒ f k v → Collect (k, v)

data Exp t where

The self typed GADT: Exp, that encodes the shape of Set expressions. A deep embedding. Exp is a typed Symbolic representation of queries we may ask. It allows us to introspect a query The strategy is to 1) Define Exp so all queries can be represented. 2) Define smart constructors that "parse" the surface syntax, and build a typed Exp 3) Write an evaluate function: eval:: Exp t -> t 4) "eval" can introspect the code and apply efficient domain and type specific translations 5) Use the (Iter f) class to evaluate some Exp that can benefit from its efficient nature. ===============================================================================================

Constructors

Base ∷ (Ord k, Basic f) ⇒ BaseRep f k v → f k v → Exp (f k v)

Dom ∷ Ord k ⇒ Exp (f k v) → Exp (Sett k ())

Rng ∷ (Ord k, Ord v) ⇒ Exp (f k v) → Exp (Sett v ())

DRestrict ∷ (Ord k, Iter g) ⇒ Exp (g k ()) → Exp (f k v) → Exp (f k v)

DExclude ∷ (Ord k, Iter g) ⇒ Exp (g k ()) → Exp (f k v) → Exp (f k v)

RRestrict ∷ (Ord k, Iter g, Ord v) ⇒ Exp (f k v) → Exp (g v ()) → Exp (f k v)

RExclude ∷ (Ord k, Iter g, Ord v) ⇒ Exp (f k v) → Exp (g v ()) → Exp (f k v)

Elem ∷ (Ord k, Iter g, Show k) ⇒ k → Exp (g k ()) → Exp Bool

NotElem ∷ (Ord k, Iter g, Show k) ⇒ k → Exp (g k ()) → Exp Bool

Intersect ∷ (Ord k, Iter f, Iter g) ⇒ Exp (f k v) → Exp (g k u) → Exp (Sett k ())

Subset ∷ (Ord k, Iter f, Iter g) ⇒ Exp (f k v) → Exp (g k u) → Exp Bool

UnionOverrideLeft ∷ (Show k, Show v, Ord k) ⇒ Exp (f k v) → Exp (g k v) → Exp (f k v)

UnionPlus ∷ (Ord k, Monoid n) ⇒ Exp (f k n) → Exp (f k n) → Exp (f k n)

UnionOverrideRight ∷ Ord k ⇒ Exp (f k v) → Exp (g k v) → Exp (f k v)

Singleton ∷ Ord k ⇒ k → v → Exp (Single k v)

SetSingleton ∷ Ord k ⇒ k → Exp (Single k ())

KeyEqual ∷ (Ord k, Iter f, Iter g) ⇒ Exp (f k v) → Exp (g k u) → Exp Bool

dRestrict ∷ (Ord k, Iter g) ⇒ Exp (g k ()) → Exp (f k v) → Exp (f k v)

rRestrict ∷ (Ord k, Iter g, Ord v) ⇒ Exp (f k v) → Exp (g v ()) → Exp (f k v)

dExclude ∷ (Ord k, Iter g) ⇒ Exp (g k ()) → Exp (f k v) → Exp (f k v)

rExclude ∷ (Ord k, Iter g, Ord v) ⇒ Exp (f k v) → Exp (g v ()) → Exp (f k v)

class HasExp s t | s → t where

Basic types are those that can be embedded into Exp. The HasExp class, encodes how to lift a Basic type into an Exp. The function toExp will build a typed Exp for that Basic type. This will be really usefull in the smart constructors. ==================================================================

Methods

toExp ∷ s → Exp t

dom ∷ (Ord k, HasExp s (f k v)) ⇒ s → Exp (Sett k ())

rng ∷ (Ord k, Ord v) ⇒ HasExp s (f k v) ⇒ s → Exp (Sett v ())

(◁) ∷ (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) ⇒ s1 → s2 → Exp (f k v)

(<|) ∷ (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) ⇒ s1 → s2 → Exp (f k v)

drestrict ∷ (Ord k, HasExp s1 (Sett k ()), HasExp s2 (f k v)) ⇒ s1 → s2 → Exp (f k v)

(⋪) ∷ (Ord k, Iter g, HasExp s1 (g k ()), HasExp s2 (f k v)) ⇒ s1 → s2 → Exp (f k v)

dexclude ∷ (Ord k, Iter g, HasExp s1 (g k ()), HasExp s2 (f k v)) ⇒ s1 → s2 → Exp (f k v)

(▷) ∷ (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) ⇒ s1 → s2 → Exp (f k v)

(|>) ∷ (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) ⇒ s1 → s2 → Exp (f k v)

rrestrict ∷ (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) ⇒ s1 → s2 → Exp (f k v)

(⋫) ∷ (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) ⇒ s1 → s2 → Exp (f k v)

rexclude ∷ (Ord k, Iter g, Ord v, HasExp s1 (f k v), HasExp s2 (g v ())) ⇒ s1 → s2 → Exp (f k v)

(∈) ∷ (Show k, Ord k, Iter g, HasExp s (g k ())) ⇒ k → s → Exp Bool

(∉) ∷ (Show k, Ord k, Iter g, HasExp s (g k ())) ⇒ k → s → Exp Bool

notelem ∷ (Show k, Ord k, Iter g, HasExp s (g k ())) ⇒ k → s → Exp Bool

(∪) ∷ (Show k, Show v, Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) ⇒ s1 → s2 → Exp (f k v)

unionleft ∷ (Show k, Show v, Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) ⇒ s1 → s2 → Exp (f k v)

(⨃) ∷ (Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) ⇒ s1 → s2 → Exp (f k v)

unionright ∷ (Ord k, HasExp s1 (f k v), HasExp s2 (g k v)) ⇒ s1 → s2 → Exp (f k v)

(∪+) ∷ (Ord k, Monoid n, HasExp s1 (f k n), HasExp s2 (f k n)) ⇒ s1 → s2 → Exp (f k n)

unionplus ∷ (Ord k, Monoid n, HasExp s1 (f k n), HasExp s2 (f k n)) ⇒ s1 → s2 → Exp (f k n)

singleton ∷ Ord k ⇒ k → v → Exp (Single k v)

setSingleton ∷ Ord k ⇒ k → Exp (Single k ())

(∩) ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp (Sett k ())

intersect ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp (Sett k ())

(⊆) ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp Bool

subset ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp Bool

(≍) ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp Bool

keyeq ∷ (Ord k, Iter f, Iter g, HasExp s1 (f k v), HasExp s2 (g k u)) ⇒ s1 → s2 → Exp Bool

data Pat env t where

Symbolc functions (Fun) are data, that can be pattern matched over. They 1) Represent a wide class of binary functions that are used in translating the SetAlgebra 2) Turned into a String so they can be printed 3) Turned into the function they represent. 4) Composed into bigger functions 5) Symbolically symplified Here we implement Symbolic Binary functions with upto 4 variables, which is enough for this use =================================================================================================

Constructors

P1 ∷ Pat (d, c, b, a) d

P2 ∷ Pat (d, c, b, a) c

P3 ∷ Pat (d, c, b, a) b

P4 ∷ Pat (d, c, b, a) a

PPair ∷ Pat (d, c, b, a) a → Pat (d, c, b, a) b → Pat (d, c, b, a) (a, b)

data Expr env t where

Constructors

X1 ∷ Expr (d, c, b, a) d

X2 ∷ Expr (d, c, b, a) c

X3 ∷ Expr (d, c, b, a) b

X4 ∷ Expr (d, c, b, a) a

HasKey ∷ (Iter f, Ord k) ⇒ Expr e k → f k v → Expr e Bool

Neg ∷ Expr e Bool → Expr e Bool

Ap ∷ Lam (a → b → c) → Expr e a → Expr e b → Expr e c

EPair ∷ Expr e a → Expr e b → Expr e (a, b)

FST ∷ Expr e (a, b) → Expr e a

SND ∷ Expr e (a, b) → Expr e b

Lit ∷ Show t ⇒ t → Expr env t

data Lam t where

Constructors

Lam ∷ Pat (d, c, b, a) t → Pat (d, c, b, a) s → Expr (d, c, b, a) v → Lam (t → s → v)

Add ∷ Num n ⇒ Lam (n → n → n)

Cat ∷ Monoid m ⇒ Lam (m → m → m)

Eql ∷ Eq t ⇒ Lam (t → t → Bool)

Both ∷ Lam (Bool → Bool → Bool)

Lift ∷ (a → b → c) → Lam (a → b → c)

type StringEnv = (String, String, String, String)

bindE ∷ Pat (a, b, c, d) t → Expr (w, x, y, z) t → StringEnv → StringEnv

showE ∷ StringEnv → Expr (a, b, c, d) t → String

showL ∷ StringEnv → Lam t → String

showP ∷ StringEnv → Pat any t → String

data Fun t

Constructors

Fun (Lam t) t

apply ∷ Fun t → t

first ∷ Fun (v → s → v)

second ∷ Fun (v → s → s)

plus ∷ Monoid t ⇒ Fun (t → t → t)

eql ∷ Eq t ⇒ Fun (t → t → Bool)

constant ∷ Show c ⇒ c → Fun (a → b → c)

rngElem ∷ (Ord rng, Iter f) ⇒ f rng v → Fun (dom → rng → Bool)

domElem ∷ (Ord dom, Iter f) ⇒ f dom v → Fun (dom → rng → Bool)

rngFst ∷ Fun (x → (a, b) → a)

rngSnd ∷ Fun (x → (a, b) → b)

compose1 ∷ Fun (t1 → t2 → t3) → Fun (t1 → t4 → t2) → Fun (t1 → t4 → t3)

compSndL ∷ Fun (k → (a, b) → c) → Fun (k → d → a) → Fun (k → (d, b) → c)

compSndR ∷ Fun (k → (a, b) → c) → Fun (k → d → b) → Fun (k → (a, d) → c)

compCurryR ∷ Fun (k → (a, b) → d) → Fun (a → c → b) → Fun (k → (a, c) → d)

nEgate ∷ Fun (k → v → Bool) → Fun (k → v → Bool)

always ∷ Fun (a → b → Bool)

both ∷ Fun (a → b → Bool) → Fun (a → b → Bool) → Fun (a → b → Bool)

lift ∷ (a → b → c) → Fun (a → b → c)

materialize ∷ Ord k ⇒ BaseRep f k v → Collect (k, v) → f k v

Given a BaseRep we can materialize a (Collect k v) into the type witnessed by the BaseRep. Recall a (Collect k v) has no intrinsic type (it is just an ABSTRACT sequence of tuples), so the witness describes how to turn them into the chosen datatype. Note that materialize is meant to be applied to a collection built by iterating over a Query. This produces the keys in ascending order, with no duplicate keys. So we do not need to specify how to merge values. =============================================================================================

addp ∷ (Ord k, Basic f) ⇒ (v → v → v) → (k, v) → f k v → f k v

fromList ∷ Ord k ⇒ BaseRep f k v → (v → v → v) → [(k, v)] → f k v

(⨝) ∷ (Ord k, Iter f, Iter g) ⇒ f k b → g k c → Collect (k, b, c)

domEq ∷ (Ord k, Iter f, Iter g) ⇒ f k b → g k c → Collect (k, b, c)

domEqSlow ∷ (Ord k, Iter f, Iter g) ⇒ f k b → g k c → Collect (k, b, c)

data Query k v where

Constructors

BaseD ∷ (Iter f, Ord k) ⇒ BaseRep f k v → f k v → Query k v

ProjectD ∷ Ord k ⇒ Query k v → Fun (k → v → u) → Query k u

AndD ∷ Ord k ⇒ Query k v → Query k w → Query k (v, w)

ChainD ∷ (Ord k, Ord v) ⇒ Query k v → Query v w → Fun (k → (v, w) → u) → Query k u

AndPD ∷ Ord k ⇒ Query k v → Query k u → Fun (k → (v, u) → w) → Query k w

OrD ∷ Ord k ⇒ Query k v → Query k v → Fun (v → v → v) → Query k v

GuardD ∷ Ord k ⇒ Query k v → Fun (k → v → Bool) → Query k v

DiffD ∷ Ord k ⇒ Query k v → Query k u → Query k v

smart ∷ Bool

projD ∷ Ord k ⇒ Query k v → Fun (k → v → u) → Query k u

andD ∷ Ord k ⇒ Query k v1 → Query k v2 → Query k (v1, v2)

andPD ∷ Ord k ⇒ Query k v1 → Query k u → Fun (k → (v1, u) → v) → Query k v

chainD ∷ (Ord k, Ord v) ⇒ Query k v → Query v w → Fun (k → (v, w) → u) → Query k u

guardD ∷ Ord k ⇒ Query k v → Fun (k → v → Bool) → Query k v

compileSubterm ∷ Exp a → Exp (f k v) → Query k v

Compile the (Exp (f k v)) to a Query iterator, and a BaseRep that indicates how to materialize the iterator to the correct type. Recall the iterator can be used to constuct many things using runCollect, but here we want to materialize it to the same type as the (Exp (f k v)), i.e. (f k v). ================================================================================

compile ∷ Exp (f k v) → (Query k v, BaseRep f k v)

testing ∷ Bool

runSetExp ∷ Ord k ⇒ Exp (f k v) → f k v

runSet ∷ Ord k ⇒ Exp (f k v) → f k v

run ∷ Ord k ⇒ (Query k v, BaseRep f k v) → f k v

runBoolExp ∷ Exp Bool → Bool

runBool ∷ Exp Bool → Bool

compute ∷ Exp t → t

eval ∷ Embed s t ⇒ Exp t → s

projStep ∷ Ord k ⇒ (t → Collect (k, v, Query k v)) → Fun (k → v → u) → t → Collect (k, u, Query k u)

andStep ∷ Ord a ⇒ (a, b1, Query a b1) → (a, b2, Query a b2) → Collect (a, (b1, b2), Query a (b1, b2))

chainStep ∷ (Ord b, Ord a) ⇒ (a, b, Query a b) → Query b w → Fun (a → (b, w) → u) → Collect (a, u, Query a u)

andPstep ∷ Ord a ⇒ (a, b1, Query a b1) → (a, b2, Query a b2) → Fun (a → (b1, b2) → w) → Collect (a, w, Query a w)

orStep ∷ (Ord k, Ord a) ⇒ (Query k v → Collect (a, v, Query k v)) → Query k v → Query k v → Fun (v → v → v) → Collect (a, v, Query k v)

guardStep ∷ Ord a ⇒ (Query a b → Collect (a, b, Query a b)) → Fun (a → b → Bool) → Query a b → Collect (a, b, Query a b)

diffStep ∷ Ord k ⇒ (k, v, Query k v) → Query k u → Collect (k, v, Query k v)

rngStep ∷ Ord v ⇒ Query k v → Sett v ()

nxtQuery ∷ Query a b → Collect (a, b, Query a b)

lubQuery ∷ Ord a ⇒ a → Query a b → Collect (a, b, Query a b)

projectQ ∷ (Ord k, HasQuery c k v) ⇒ c → Fun (k → v → u) → Query k u

andQ ∷ (Ord k, HasQuery concrete1 k v, HasQuery concrete2 k w) ⇒ concrete1 → concrete2 → Query k (v, w)

orQ ∷ (Ord k, HasQuery concrete1 k v, HasQuery concrete2 k v) ⇒ concrete1 → concrete2 → Fun (v → v → v) → Query k v

chainQ ∷ (Ord k, Ord v, HasQuery concrete1 k v, HasQuery concrete2 v w) ⇒ concrete1 → concrete2 → Fun (k → (v, w) → u) → Query k u

andPQ ∷ (Ord k, HasQuery concrete1 k v, HasQuery concrete2 k u) ⇒ concrete1 → concrete2 → Fun (k → (v, u) → w) → Query k w

guardQ ∷ (Ord k, HasQuery concrete k v) ⇒ concrete → Fun (k → v → Bool) → Query k v

class HasQuery concrete k v where

Methods

query ∷ concrete → Query k v

ppQuery ∷ Query k v → Doc