Control.State.Transition.Extended

Description

Small step state transition systems.

Documentation

data RuleType

Constructors

Initial

Transition

class RuleTypeRep t

Minimal complete definition

rTypeRep

type family RuleContext (t ∷ RuleType) = (ctx ∷ Type → Type) | ctx → t where ...

Equations

RuleContext Initial = IRC

RuleContext Transition = TRC

newtype IRC sts

Context available to initial rules.

Constructors

IRC (Environment sts)

newtype TRC sts

Context available to transition rules.

Constructors

TRC (Environment sts, State sts, Signal sts)

type Rule sts rtype = F (Clause sts rtype)

type TransitionRule sts = Rule sts Transition (State sts)

type InitialRule sts = Rule sts Initial (State sts)

data Assertion sts

An assertion is a validation condition for the STS system in question. It should be used to define properties of the system as a whole that cannot be violated under normal circumstances - e.g. a violation implies a failing in the rule logic.

Assertions should not check for conditions that may differ between different rules in a system, since the interpreter may stop the system upon presence of a failed assertion.

Whether assertions are checked is a matter for the STS interpreter.

Constructors

PreCondition String (TRC sts → Bool)
Pre-condition. Checked before the rule fires.

PostCondition String (TRC sts → State sts → Bool)
Post-condition. Checked after the rule fires, and given access to the resultant state as well as the initial context.

data AssertionViolation sts

Constructors

AssertionViolation

Fields

avSTS ∷ String

avMsg ∷ String

avCtx ∷ TRC sts

avState ∷ Maybe (State sts)

class (Eq (PredicateFailure a), Show (PredicateFailure a), Monad (BaseM a), Typeable a) ⇒ STS a where

State transition system.

Minimal complete definition

initialRules, transitionRules

Associated Types

type State a ∷ Type

Type of the state which the system transitions between.

type Signal a ∷ Type

Signal triggering a state change.

type Environment a ∷ Type

Environment type.

type BaseM a ∷ Type → Type

Monad into which to interpret the rules.

type PredicateFailure a = (b ∷ Type) | b → a

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Methods

initialRules ∷ [InitialRule a]

Rules governing transition under this system.

transitionRules ∷ [TransitionRule a]

assertions ∷ [Assertion a]

Assertions about the transition system.

renderAssertionViolation ∷ AssertionViolation a → String

Render an assertion violation.

Defaults to using show, but note that this does not know how to render the context. So for more information you should define your own renderer here.

class (STS sub, STS super, BaseM sub ~ BaseM super) ⇒ Embed sub super where

Embed one STS within another.

Methods

wrapFailed ∷ PredicateFailure sub → PredicateFailure super

Wrap a predicate failure of the subsystem in a failure of the super-system.

(?!) ∷ Bool → PredicateFailure sts → Rule sts ctx () infix 1

Oh noes!

This takes a condition (a boolean expression) and a failure and results in a clause which will throw that failure if the condition fails.

(?!:) ∷ Either e () → (e → PredicateFailure sts) → Rule sts ctx ()

Oh noes with an explanation

We interpret this as "What?" "No!" "Because:"

failBecause ∷ PredicateFailure sts → Rule sts ctx ()

judgmentContext ∷ Rule sts rtype (RuleContext rtype sts)

Get the judgment context

trans ∷ Embed sub super ⇒ RuleContext rtype sub → Rule super rtype (State sub)

liftSTS ∷ STS sts ⇒ BaseM sts a → Rule sts ctx a

data AssertionPolicy

Control which assertions are enabled.

Constructors

AssertionsAll

AssertionsPre
Only run preconditions

AssertionsPost
Only run postconditions

AssertionsOff

data ValidationPolicy

Control which predicates are evaluated during rule processing.

Constructors

ValidateAll

ValidateNone

data ApplySTSOpts

Constructors

ApplySTSOpts

Fields

asoAssertions ∷ AssertionPolicy

Enable assertions during STS processing. If this option is enabled, STS processing will terminate on violation of an assertion.

asoValidation ∷ ValidationPolicy

Validation policy

applySTSOpts ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ ApplySTSOpts → RuleContext rtype s → m (State s, [[PredicateFailure s]])

Apply an STS with options. Note that this returns both the final state and the list of predicate failures.

applySTS ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (Either [[PredicateFailure s]] (State s))

applySTSIndifferently ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (State s, [[PredicateFailure s]])

reapplySTS ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (State s)

Re-apply an STS.

It is assumed that the caller of this function has previously applied this STS, and can guarantee that it completed successfully. No predicates will be checked when calling this function.

Random thing

newtype Threshold a

This can be used to specify predicate failures in STS rules where a value is beyond a certain threshold.

TODO move this somewhere more sensible

Constructors

Threshold a