Data.CannonicalMaps

Documentation

class Eq t ⇒ CannonicalZero t where

Methods

zeroC ∷ t

joinC ∷ t → t → t

cannonicalInsert ∷ (Ord k, CannonicalZero a) ⇒ (a → a → a) → k → a → Map k a → Map k a

cannonicalMapUnion ∷ (Ord k, CannonicalZero a) ⇒ (a → a → a) → Map k a → Map k a → Map k a

cannonicalMap ∷ (Ord k, CannonicalZero a) ⇒ (a → a) → Map k a → Map k a

pointWise ∷ (Ord k, CannonicalZero v) ⇒ (v → v → Bool) → Map k v → Map k v → Bool

data Map k a

A Map from keys k to values a.

The Semigroup operation for Map is union, which prefers values from the left operand. If m1 maps a key k to a value a1, and m2 maps the same key to a different value a2, then their union m1 <> m2 maps k to a1.