Hedgehog.Extra.Manual

Description

Manual generators.

This module provides functions to convert hedgehog Gens to and from a Manual generators, and functions to manipulate these manual generators.

Documentation

newtype Manual a

Constructors

Manual (Size → Seed → a)

unManual ∷ Manual a → Size → Seed → a

toManual ∷ Gen a → Manual (TreeT (MaybeT Identity) a)

fromManual ∷ Manual (TreeT (MaybeT Identity) a) → Gen a

dontShrink ∷ Gen a → Manual (Maybe a)

Combinators

sized ∷ (Size → Manual a) → Manual a

replicate ∷ ∀ a. Int → Manual a → Manual [a]

A version of replicateM specific to Manual.

interleave ∷ [TreeT (MaybeT Identity) a] → TreeT (MaybeT Identity) [a]

Auxiliary

wrapTreeT ∷ Maybe (NodeT (MaybeT Identity) a) → TreeT (MaybeT Identity) a

unwrapTreeT ∷ TreeT (MaybeT Identity) a → Maybe (NodeT (MaybeT Identity) a)