Control.State.Transition.Extended

Description

Small step state transition systems.

Documentation

data RuleType

Constructors

Initial

Transition

class RuleTypeRep t

Minimal complete definition

rTypeRep

type family RuleContext (t ∷ RuleType) = (ctx ∷ Type → Type) | ctx → t where ...

Equations

RuleContext 'Initial = IRC

RuleContext 'Transition = TRC

newtype IRC sts

Context available to initial rules.

Constructors

IRC (Environment sts)

newtype TRC sts

Context available to transition rules.

Constructors

TRC (Environment sts, State sts, Signal sts)

type Rule sts rtype = F (Clause sts rtype)

type TransitionRule sts = Rule sts 'Transition (State sts)

type InitialRule sts = Rule sts 'Initial (State sts)

data Assertion sts

An assertion is a validation condition for the STS system in question. It should be used to define properties of the system as a whole that cannot be violated under normal circumstances - e.g. a violation implies a failing in the rule logic.

Assertions should not check for conditions that may differ between different rules in a system, since the interpreter may stop the system upon presence of a failed assertion.

Whether assertions are checked is a matter for the STS interpreter.

Constructors

PreCondition String (TRC sts → Bool)
Pre-condition. Checked before the rule fires.

PostCondition String (TRC sts → State sts → Bool)
Post-condition. Checked after the rule fires, and given access to the resultant state as well as the initial context.

data AssertionViolation sts

Constructors

AssertionViolation

Fields

avSTS ∷ String

avMsg ∷ String

avCtx ∷ TRC sts

avState ∷ Maybe (State sts)

data AssertionException where

Constructors

AssertionException ∷ ∀ sts. STS sts ⇒ AssertionViolation sts → AssertionException

class (Eq (PredicateFailure a), Show (PredicateFailure a), Monad (BaseM a), Typeable a) ⇒ STS a where

State transition system.

Minimal complete definition

transitionRules

Associated Types

type State a ∷ Type

Type of the state which the system transitions between.

type Signal a ∷ Type

Signal triggering a state change.

type Environment a ∷ Type

Environment type.

type BaseM a ∷ Type → Type

Monad into which to interpret the rules.

type BaseM a = Identity

type Event a ∷ Type

Event type.

type Event a = Void

type PredicateFailure a ∷ Type

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Methods

initialRules ∷ [InitialRule a]

Rules governing transition under this system.

default initialRules ∷ Default (State a) ⇒ [InitialRule a]

transitionRules ∷ [TransitionRule a]

assertions ∷ [Assertion a]

Assertions about the transition system.

renderAssertionViolation ∷ AssertionViolation a → String

Render an assertion violation.

Defaults to using show, but note that this does not know how to render the context. So for more information you should define your own renderer here.

data STUB (e ∷ Type) (st ∷ Type) (si ∷ Type) (f ∷ Type) (m ∷ Type → Type)

A stub rule with no transitions to use as a placeholder

class (STS sub, BaseM sub ~ BaseM super) ⇒ Embed sub super where

Embed one STS within another.

Minimal complete definition

wrapFailed

Methods

wrapFailed ∷ PredicateFailure sub → PredicateFailure super

Wrap a predicate failure of the subsystem in a failure of the super-system.

wrapEvent ∷ Event sub → Event super

default wrapEvent ∷ Coercible (Event sub) (Event super) ⇒ Event sub → Event super

(?!) ∷ Bool → PredicateFailure sts → Rule sts ctx () infix 1

Oh noes!

This takes a condition (a boolean expression) and a failure and results in a clause which will throw that failure if the condition fails.

(?!:) ∷ Either e () → (e → PredicateFailure sts) → Rule sts ctx ()

Oh noes with an explanation

We interpret this as "What?" "No!" "Because:"

validate ∷ Validation (NonEmpty (PredicateFailure sts)) () → Rule sts ctx ()

Fail with PredicateFailure's in STS if Validation was unsuccessful.

validateTrans

Arguments

∷ (e → PredicateFailure sts)
Transformation function for all errors

→ Validation (NonEmpty e) ()

→ Rule sts ctx ()

Same as validation, except with ability to transform opaque failures into PredicateFailures with a help of supplied function.

validateTransLabeled

Arguments

∷ (e → PredicateFailure sts)
Transformation function for all errors

→ NonEmpty Label
Supply a list of labels to be used as filters when STS is executed

→ Validation (NonEmpty e) ()
Actual validations to be executed

→ Rule sts ctx ()

Same as validation, except with ability to translate opaque failures into PredicateFailures with a help of supplied function.

type Label = String

Label for a predicate. This can be used to control which predicates get run.

data SingEP ep where

Constructors

EPReturn ∷ SingEP 'EventPolicyReturn

EPDiscard ∷ SingEP 'EventPolicyDiscard

data EventPolicy

Constructors

EventPolicyReturn

EventPolicyDiscard

type family EventReturnType ep sts a ∷ Type where ...

Equations

EventReturnType 'EventPolicyReturn sts a = (a, [Event sts])

EventReturnType _ _ a = a

labeled ∷ NonEmpty Label → Rule sts ctx () → Rule sts ctx ()

Labeled clause. This will only be executed if the interpreter is set to execute clauses with this label.

labeledPred ∷ NonEmpty Label → Bool → PredicateFailure sts → Rule sts ctx ()

Labeled predicate. This may be used to control which predicates are run using ValidateSuchThat.

labeledPredE ∷ NonEmpty Label → Either e () → (e → PredicateFailure sts) → Rule sts ctx ()

Labeled predicate with an explanation

ifFailureFree ∷ Rule sts rtype a → Rule sts rtype a → Rule sts rtype a

whenFailureFree ∷ Rule sts rtype () → Rule sts rtype ()

failBecause ∷ PredicateFailure sts → Rule sts ctx ()

failOnJust ∷ Maybe a → (a → PredicateFailure sts) → Rule sts ctx ()

Produce a predicate failure when condition contains a Just value, contents of which can be used inside the predicate failure

failOnNonEmpty ∷ Foldable f ⇒ f a → (NonEmpty a → PredicateFailure sts) → Rule sts ctx ()

Produce a predicate failure when supplied foldable is not empty, contents of which will be converted to a NonEmpty list and can be used inside the predicate failure.

failureOnJust ∷ Maybe a → (a → e) → Validation (NonEmpty e) ()

Produce a failure when condition contains a Just value, contents of which can be used inside the failure

failureOnNonEmpty ∷ Foldable f ⇒ f a → (NonEmpty a → e) → Validation (NonEmpty e) ()

Produce a failure when supplied foldable is not empty, contents of which will be converted to a NonEmpty list and theh can be used inside the failure.

judgmentContext ∷ Rule sts rtype (RuleContext rtype sts)

Get the judgment context

trans ∷ Embed sub super ⇒ RuleContext rtype sub → Rule super rtype (State sub)

liftSTS ∷ STS sts ⇒ BaseM sts a → Rule sts ctx a

tellEvent ∷ Event sts → Rule sts ctx ()

tellEvents ∷ [Event sts] → Rule sts ctx ()

class EventReturnTypeRep ert

Minimal complete definition

eventReturnTypeRep

mapEventReturn ∷ ∀ ep sts a b. EventReturnTypeRep ep ⇒ (a → b) → EventReturnType ep sts a → EventReturnType ep sts b

Map over an arbitrary EventReturnType.

Apply STS

data AssertionPolicy

Control which assertions are enabled.

Constructors

AssertionsAll

AssertionsPre
Only run preconditions

AssertionsPost
Only run postconditions

AssertionsOff

data ValidationPolicy

Control which predicates are evaluated during rule processing.

Constructors

ValidateAll

ValidateNone

ValidateSuchThat ([Label] → Bool)

data ApplySTSOpts ep

Constructors

ApplySTSOpts

Fields

asoAssertions ∷ AssertionPolicy

Enable assertions during STS processing. If this option is enabled, STS processing will terminate on violation of an assertion.

asoValidation ∷ ValidationPolicy

Validation policy

asoEvents ∷ SingEP ep

Event policy

applySTSOpts ∷ ∀ s m rtype ep. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ ApplySTSOpts ep → RuleContext rtype s → m (EventReturnType ep s (State s, [PredicateFailure s]))

Apply an STS with options. Note that this returns both the final state and the list of predicate failures.

applySTSOptsEither ∷ ∀ s m rtype ep. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ ApplySTSOpts ep → RuleContext rtype s → m (Either (NonEmpty (PredicateFailure s)) (EventReturnType ep s (State s)))

applySTS ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (Either (NonEmpty (PredicateFailure s)) (State s))

applySTSIndifferently ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (State s, [PredicateFailure s])

reapplySTS ∷ ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (State s)

Re-apply an STS.

It is assumed that the caller of this function has previously applied this STS, and can guarantee that it completed successfully. No predicates or assertions will be checked when calling this function.

globalAssertionPolicy ∷ AssertionPolicy

Exported to allow running rules independently

applySTSInternal

Arguments

∷ ∀ s m rtype ep. (STS s, RuleTypeRep rtype, m ~ BaseM s)

⇒ SingEP ep

→ AssertionPolicy

→ RuleInterpreter ep
Interpreter for rules

→ RuleContext rtype s

→ ExceptT (AssertionViolation s) m (EventReturnType ep s (State s, [PredicateFailure s]))

applyRuleInternal

Arguments

∷ ∀ (ep ∷ EventPolicy) s m rtype result. (STS s, RuleTypeRep rtype, m ~ BaseM s)

⇒ IsFailing
We need to know if the current STS incurred at least one PredicateFailure. This is necessary because applyRuleInternal is called recursively through the goSTS argument, which will not have access to any of the predicate failures occured in other branches of STS rule execusion tree.

→ SingEP ep

→ ValidationPolicy

→ (IsFailing → STSInterpreter ep)
Interpreter for subsystems

→ RuleContext rtype s

→ Rule s rtype result

→ m (EventReturnType ep s (result, [PredicateFailure s]))

Apply a rule even if its predicates fail.

If the rule successfully applied, the list of predicate failures will be empty.

type RuleInterpreter ep = ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → Rule s rtype (State s) → m (EventReturnType ep s (State s, [PredicateFailure s]))

type STSInterpreter ep = ∀ s m rtype. (STS s, RuleTypeRep rtype, m ~ BaseM s) ⇒ RuleContext rtype s → m (EventReturnType ep s (State s, [PredicateFailure s]))

runRule ∷ (STS s, RuleTypeRep rtype) ⇒ RuleContext rtype s → Rule s rtype result → BaseM s (result, [PredicateFailure s])

runRule :: RuleInterpreter 'EventPolicyDiscard run a rule given its context, to get a BaseM computation

Random thing

newtype Threshold a

This can be used to specify predicate failures in STS rules where a value is beyond a certain threshold.

TODO move this somewhere more sensible

Constructors

Threshold a

sfor_ ∷ (Foldable t, Applicative f) ⇒ t a → (a → f b) → f ()

sfor_ is straverse_ with its arguments flipped. For a version that doesn't ignore the results see for.

>>> sfor_ ([1..4] :: [Int]) print
1
2
3
4