Control.State.Transition.Simple

Description

Simple state transition system over the Identity monad.

Documentation

applyRuleIndifferently ∷ ∀ s rtype. (STS s, RuleTypeRep rtype, Identity ~ BaseM s) ⇒ RuleContext rtype s → Rule s rtype (State s) → (State s, [PredicateFailure s])

Apply a rule even if its predicates fail.

If the rule successfully applied, the list of predicate failures will be empty.

applySTSIndifferently ∷ ∀ s rtype. (STS s, RuleTypeRep rtype, Identity ~ BaseM s) ⇒ RuleContext rtype s → (State s, [[PredicateFailure s]])

applySTS ∷ ∀ s rtype. (STS s, RuleTypeRep rtype, BaseM s ~ Identity) ⇒ RuleContext rtype s → Either [[PredicateFailure s]] (State s)

newtype Threshold a

This can be used to specify predicate failures in STS rules where a value is beyond a certain threshold.

Constructors

Threshold a

type Rule sts rtype = F (Clause sts rtype)

data Clause sts (rtype ∷ RuleType) a where

Constructors

Lift ∷ STS sts ⇒ BaseM sts a → (a → b) → Clause sts rtype b

GetCtx ∷ (RuleContext rtype sts → a) → Clause sts rtype a

SubTrans ∷ Embed sub sts ⇒ RuleContext rtype sub → (State sub → a) → Clause sts rtype a

Predicate ∷ Either e a → (e → PredicateFailure sts) → a → Clause sts rtype a

class (STS sub, STS super, BaseM sub ~ BaseM super) ⇒ Embed sub super where

Embed one STS within another.

Methods

wrapFailed ∷ PredicateFailure sub → PredicateFailure super

Wrap a predicate failure of the subsystem in a failure of the super-system.

class (Eq (PredicateFailure a), Show (PredicateFailure a), Monad (BaseM a)) ⇒ STS a where

State transition system.

Associated Types

type State a ∷ ★

Type of the state which the system transitions between.

type Signal a ∷ ★

Signal triggering a state change.

type Environment a ∷ ★

Environment type.

type BaseM a ∷ ★ → ★

Monad into which to interpret the rules.

data PredicateFailure a ∷ ★

Descriptive type for the possible failures which might cause a transition to fail.

As a convention, PredicateFailures which are "structural" (meaning that they are not "throwable" in practice, and are used to pass control from one transition rule to another) are prefixed with S_.

Structural PredicateFailures represent conditions between rules where the disjunction of all rules' preconditions is equal to True. That is, either one rule will throw a structural PredicateFailure and the other will succeed, or vice-versa.

Methods

initialRules ∷ [InitialRule a]

Rules governing transition under this system.

transitionRules ∷ [TransitionRule a]

type TransitionRule sts = Rule sts Transition (State sts)

type InitialRule sts = Rule sts Initial (State sts)

type family RuleContext (t ∷ RuleType) = (ctx ∷ Type → Type) | ctx → t where ...

Equations

RuleContext Initial = IRC

RuleContext Transition = TRC

newtype TRC sts

Context available to transition rules.

Constructors

TRC (Environment sts, State sts, Signal sts)

newtype IRC sts

Context available to initial rules.

Constructors

IRC (Environment sts)

class RuleTypeRep t where

Methods

rTypeRep ∷ SRuleType t

data SRuleType a where

Singleton instances.

Since our case is so small we don't bother with the singletons library.

Constructors

SInitial ∷ SRuleType Initial

STransition ∷ SRuleType Transition

data RuleType

Constructors

Initial

Transition

(?!) ∷ Bool → PredicateFailure sts → Rule sts ctx () infix 1

Oh noes!

This takes a condition (a boolean expression) and a failure and results in a clause which will throw that failure if the condition fails.

failBecause ∷ PredicateFailure sts → Rule sts ctx ()

(?!:) ∷ Either e () → (e → PredicateFailure sts) → Rule sts ctx ()

Oh noes with an explanation

We interpret this as "What?" "No!" "Because:"

trans ∷ Embed sub super ⇒ RuleContext rtype sub → Rule super rtype (State sub)

liftSTS ∷ STS sts ⇒ BaseM sts a → Rule sts ctx a

judgmentContext ∷ Rule sts rtype (RuleContext rtype sts)

Get the judgment context