Data.Coders

Description

MemoBytes is an abstration for a datetype that encodes its own seriialization. The idea is to use a newtype around a MemoBytes non-memoizing version. For example: newtype Foo = Foo(MemoBytes NonMemoizingFoo) This way all the instances for Foo (Eq,Show,Ord,ToCBOR,FromCBOR,NoThunks,Generic) can be derived for free.

Documentation

data Encode (w ∷ Wrapped) t where

Constructors

Rec ∷ t → Encode ('Closed 'Dense) t

Sum ∷ t → Word → Encode 'Open t

Keyed ∷ t → Encode ('Closed 'Sparse) t

To ∷ ToCBOR a ⇒ a → Encode ('Closed 'Dense) a

E ∷ (t → Encoding) → t → Encode ('Closed 'Dense) t

ED ∷ Dual t → t → Encode ('Closed 'Dense) t

OmitC ∷ t → Encode w t

Omit ∷ (t → Bool) → Encode ('Closed 'Sparse) t → Encode ('Closed 'Sparse) t

Key ∷ Word → Encode ('Closed 'Dense) t → Encode ('Closed 'Sparse) t

ApplyE ∷ Encode w (a → t) → Encode ('Closed r) a → Encode w t

data Decode (w ∷ Wrapped) t where

Constructors

Summands ∷ String → (Word → Decode 'Open t) → Decode ('Closed 'Dense) t

SparseKeyed ∷ Typeable t ⇒ String → t → (Word → Field t) → [(Word, String)] → Decode ('Closed 'Dense) t

SumD ∷ t → Decode 'Open t

RecD ∷ t → Decode ('Closed 'Dense) t

KeyedD ∷ t → Decode ('Closed 'Sparse) t

From ∷ FromCBOR t ⇒ Decode w t

D ∷ (∀ s. Decoder s t) → Decode ('Closed 'Dense) t

ApplyD ∷ Decode w1 (a → t) → Decode ('Closed d) a → Decode w1 t

Invalid ∷ Word → Decode w t

Map ∷ (a → b) → Decode w a → Decode w b

DD ∷ Dual t → Decode ('Closed 'Dense) t

Emit ∷ t → Decode w t

Ann ∷ Decode w t → Decode w (Annotator t)

ApplyAnn ∷ Decode w1 (Annotator (a → t)) → Decode ('Closed d) (Annotator a) → Decode w1 (Annotator t)

(!>) ∷ Encode w (a → t) → Encode ('Closed r) a → Encode w t infixl 4

(<!) ∷ Decode w1 (a → t) → Decode ('Closed w) a → Decode w1 t infixl 4

(<*!) ∷ Decode w1 (Annotator (a → t)) → Decode ('Closed d) (Annotator a) → Decode w1 (Annotator t) infixl 4

data Density

Some CBOR instances wrap encoding sequences with prefixes and suffixes. I.e. prefix , encode, encode, encode , ... , suffix. There are two kinds of wrapping coders: Nary sums, and Sparsely encoded products. Coders in these classes can only be decoded when they are wrapped by their closing forms Summand and SparseKeyed. In another dimension records can be encoded densely (all their fields serialised) or sparsely (only some of their fields). We use indexes to types to try and mark (and enforce) these distinctions.

Record density (all the fields) vs (some of the fields)

Constructors

Dense

Sparse

data Wrapped where

Constructors

Open ∷ Wrapped

Closed ∷ Density → Wrapped

newtype Annotator a

A value of type `Annotator a` is one that needs access to the entire | bytestring used during decoding to finish construction.

Constructors

Annotator

Fields

runAnnotator ∷ FullByteString → a

data Dual t

Analogous to paired ToCBOR and FromCBOR instances with out freezing out alternate ways to code. Unlike ToCBOR and FromCBOR where there is only one instance per type. There can be multiple Duals with the same type.

Constructors

Dual (t → Encoding) (∀ s. Decoder s t)

data Field t where

A Field pairs an update function and a decoder for one field of a Sparse record.

Constructors

Field ∷ (x → t → t) → Decode ('Closed d) x → Field t

encode ∷ Encode w t → Encoding

decode ∷ Decode w t → Decoder s t

runE ∷ Encode w t → t

decodeList ∷ Decoder s a → Decoder s [a]

decodeSeq ∷ Decoder s a → Decoder s (Seq a)

decodeStrictSeq ∷ Decoder s a → Decoder s (StrictSeq a)

decodeSet ∷ Ord a ⇒ Decoder s a → Decoder s (Set a)

decodeRecordNamed ∷ Text → (a → Int) → Decoder s a → Decoder s a

decodeRecordSum ∷ String → (Word → Decoder s (Int, a)) → Decoder s a

invalidKey ∷ Word → Decoder s a

wrapCBORArray ∷ Word → Encoding → Encoding

encodeFoldable ∷ (ToCBOR a, Foldable f) ⇒ f a → Encoding

decodeCollectionWithLen ∷ Decoder s (Maybe Int) → Decoder s a → Decoder s (Int, [a])

decodeCollection ∷ Decoder s (Maybe Int) → Decoder s a → Decoder s [a]

encodeFoldableEncoder ∷ Foldable f ⇒ (a → Encoding) → f a → Encoding

dualList ∷ (ToCBOR a, FromCBOR a) ⇒ Dual [a]

dualSeq ∷ (ToCBOR a, FromCBOR a) ⇒ Dual (Seq a)

dualSet ∷ (Ord a, ToCBOR a, FromCBOR a) ⇒ Dual (Set a)

dualMaybeAsList ∷ (ToCBOR a, FromCBOR a) ⇒ Dual (Maybe a)

Good for encoding (Maybe t) if is another Maybe. Uses more space than dualMaybeAsNull

dualMaybeAsNull ∷ (ToCBOR a, FromCBOR a) ⇒ Dual (Maybe a)

Good for encoding (Maybe T) as long as T isn't another Maybe

dualText ∷ Dual Text

dualStrictSeq ∷ (ToCBOR a, FromCBOR a) ⇒ Dual (StrictSeq a)

dualCBOR ∷ (ToCBOR a, FromCBOR a) ⇒ Dual a

to ∷ (ToCBOR t, FromCBOR t) ⇒ t → Encode ('Closed 'Dense) t

from ∷ (ToCBOR t, FromCBOR t) ⇒ Decode ('Closed 'Dense) t

data Decoder s a

data Encoding

encodeNullMaybe ∷ (a → Encoding) → Maybe a → Encoding

decodeNullMaybe ∷ Decoder s a → Decoder s (Maybe a)